+++ src/api/mod.rs
@@ -1,5 +1,5 @@
 pub mod model;
-mod api_utils;
+pub mod api_utils;
 mod scheduler;
 mod endpoints;
 pub mod main_api;